libport.allocator-static: use size_t.
	* include/libport/allocator-static.hh,
	* include/libport/allocator-static.hxx,
	* tests/libport/allocator-static.cc: Fix warnings on Mac OS X.

+++ include/libport/allocator-static.hh
@@ -11,13 +11,15 @@
 #ifndef ALLOCATOR_STATIC_HH
 # define ALLOCATOR_STATIC_HH
 
+# include <cstdlib>
+
 namespace libport
 {
-  template <unsigned Max, unsigned Size>
+  template <unsigned Max, size_t Size>
   class StaticallyAllocated
   {
   public:
-    void* operator new(unsigned int);
+    void* operator new(size_t);
     void operator delete(void* obj);
 
     static int initialize();
